add resize

+++ backend/board/consumers.py
@@ -8,6 +8,7 @@ class BoardConsumer(AsyncJsonWebsocketConsumer):
         super().__init__(*args, **kwargs)
         self.user = None
         self.board_id = None
+        self.delete_obj_basket = True
 
     async def connect(self):
         try:
@@ -86,7 +87,7 @@ class BoardConsumer(AsyncJsonWebsocketConsumer):
             await self.redo_object(content)
             await self.send_update_board_data()
         elif content.get('method') == 'resize':
-            await self.send_update_board_data()
+            self.delete_obj_basket = False
         else:
             board_obj, undo_obj = await self.add_object(content)
             await self.channel_layer.group_send(
@@ -99,4 +100,5 @@ class BoardConsumer(AsyncJsonWebsocketConsumer):
             )
 
     async def disconnect(self, code):
-        await self.delete_board_data_basket_objects()
+        if self.delete_obj_basket:
+            await self.delete_board_data_basket_objects()
